Split resources.py into three files: instances.py, resources.py, and servers.py
